+++ include/crpropa/Source.h
@@ -342,44 +342,53 @@ public:
  @brief Source distribution that follows the Galactic SNR distribution
 
  The origin of the distribution is the Galactic center. The default maximum radius is set 
- to R_max=20 kpc and the default maximum height is Z_max = 5 kpc.
+ to rMax=20 kpc and the default maximum height is zMax = 5 kpc.
  See G. Case and D. Bhattacharya (1996) for the details of the distribution.
  */
 class SourceSNRDistribution: public SourceFeature {
-	double R_earth; // parameter given by observation
+	double rEarth; // parameter given by observation
+	double alpha; // parameter to shift the maximum in R direction
 	double beta; // parameter to shift the maximum in R direction
-	double Zg; // exponential cut parameter in z direction
+	double zg; // exponential cut parameter in z direction
 	double frMax; // helper for efficient sampling
 	double fzMax; // helper for efficient sampling
-	double R_max; // maximum radial distance - default 20 kpc (due to the extension of the JF12 field)
-	double Z_max; // maximum distance from galactic plane - default 5 kpc
+	double rMax; // maximum radial distance - default 20 kpc 
+		      // (due to the extension of the JF12 field)
+	double zMax; // maximum distance from galactic plane - default 5 kpc
 public:
 	/** Default constructor. 
 	 Default parameters are:
-	 . R_earth = 8.5 kpc
+	 . rEarth = 8.5 kpc
+	 . alpha = 2
 	 . beta = 3.53
-	 . Zg = 300 pc
-	 . Rmax = 20 kpc
-	 . Zmax = 5 kpc
-	 */ 
+	 . zg = 300 pc
+	 . rMax = 20 kpc
+	 . zMax = 5 kpc
+	*/ 
 	SourceSNRDistribution();
 	/** Generic constructor
-	 @param R_earth	  distance from Earth to the Galactic centre [in meters]
+	 @param rEarth	  distance from Earth to the Galactic centre [in meters]
+	 @param alpha	  parameter that shifts radially the maximum of the distributions
 	 @param beta	  parameter that shifts radially the maximum of the distributions 
-	 @param Zg		  exponential cut-off parameter in the z-direction [in meters]
-	 */	
-	SourceSNRDistribution(double R_earth, double beta, double Zg);
+	 @param zg		  exponential cut-off parameter in the z-direction [in meters]
+	*/	
+	SourceSNRDistribution(double rEarth,double alpha, double beta, double zg);
+
 	void prepareParticle(ParticleState &particle) const;
-	double f_r(double r) const;
-	double f_z(double z) const;
-	void set_frMax(double R, double b);
-	void set_fzMax(double Zg);
-	void set_RMax(double R_max);
-	void set_ZMax(double Z_max);
-	double get_frMax();
-	double get_fzMax();
-	double get_RMax();
-	double get_ZMax();
+	double fr(double r) const;
+	double fz(double z) const;
+	void setFrMax();
+	void setFzMax(double Zg);
+	void setRMax(double rMax);
+	void setZMax(double zMax);
+	void setAlpha(double a);
+	void setBeta(double b);
+	double getFrMax() const;
+	double getFzMax() const;
+	double getRMax() const;
+	double getZMax() const;
+	double getAlpha() const;
+	double getBeta() const;
 	void setDescription();
 };
 
@@ -395,21 +404,21 @@ public:
  parametrized as in Blasi and Amato, JCAP 1 (Jan., 2012) 10.
  */
 class SourcePulsarDistribution: public SourceFeature {
-	double R_earth; // parameter given by observation
+	double rEarth; // parameter given by observation
 	double beta; // parameter to shift the maximum in R direction
-	double Zg; // exponential cut parameter in z direction
+	double zg; // exponential cut parameter in z direction
 	double frMax; // helper for efficient sampling
 	double fzMax; // helper for efficient sampling
-	double R_max; // maximum radial distance - default 22 kpc 
-	double Z_max; // maximum distance from galactic plane - default 5 kpc
-	double r_blur; // relative smearing factor for the radius
-	double theta_blur; // smearing factor for the angle. Unit = [1/length]
+	double rMax; // maximum radial distance - default 22 kpc 
+	double zMax; // maximum distance from galactic plane - default 5 kpc
+	double rBlur; // relative smearing factor for the radius
+	double thetaBlur; // smearing factor for the angle. Unit = [1/length]
 public:
 	/** Default constructor. 
 	 Default parameters are:
-	 . R_earth = 8.5 kpc
+	 . rEarth = 8.5 kpc
 	 . beta = 3.53
-	 . Zg = 300 pc
+	 . zg = 300 pc
 	 . Rmax = 22 kpc
 	 . Zmax = 5 kpc
 	 . rBlur = 0.07
@@ -417,31 +426,31 @@ public:
 	 */ 
 	SourcePulsarDistribution();	
 	/** Generic constructor
-	 @param R_earth		distance from Earth to the Galactic centre [in meters]
+	 @param rEarth		distance from Earth to the Galactic centre [in meters]
 	 @param beta		parameter that shifts radially the maximum of the distributions 
-	 @param Zg			exponential cut-off parameter in the z-direction [in meters]
-	 @param r_blur		relative smearing factor for radius
-	 @param theta_blur	smearing factor for the angle [in 1 / meters]
+	 @param zg			exponential cut-off parameter in the z-direction [in meters]
+	 @param rBlur		relative smearing factor for radius
+	 @param thetaBlur	smearing factor for the angle [in 1 / meters]
 	 */	
-	SourcePulsarDistribution(double R_earth, double beta, double Zg, double r_blur, double theta_blur);
+	SourcePulsarDistribution(double rEarth, double beta, double zg, double rBlur, double thetaBlur);
 	void prepareParticle(ParticleState &particle) const;
-	double f_r(double r) const;
-	double f_z(double z) const;
-	double f_theta(int i, double r) const;
-	double blur_r(double r_tilde) const;
-	double blur_theta(double theta_tilde, double r_tilde) const;
-	void set_frMax(double R, double b);
-	void set_fzMax(double Zg);
-	void set_RMax(double R_max);
-	void set_ZMax(double Z_max);
-	void set_rBlur(double r_blur);
-	void set_thetaBlur(double theta_blur);
-	double get_frMax();
-	double get_fzMax();
-	double get_RMax();
-	double get_ZMax();
-	double get_rBlur();
-	double get_thetaBlur();
+	double fr(double r) const;
+	double fz(double z) const;
+	double ftheta(int i, double r) const;
+	double blurR(double r_tilde) const;
+	double blurTheta(double theta_tilde, double r_tilde) const;
+	void setFrMax(double R, double b);
+	void setFzMax(double zg);
+	void setRMax(double rMax);
+	void setZMax(double zMax);
+	void setRBlur(double rBlur);
+	void setThetaBlur(double thetaBlur);
+	double getFrMax();
+	double getFzMax();
+	double getRMax();
+	double getZMax();
+	double getRBlur();
+	double getThetaBlur();
 	void setDescription();
 };
 
@@ -520,6 +529,39 @@ public:
 	void setDescription();
 };
 
+
+/**
+ @class SourceDirectedEmission
+ @brief Directed emission from a source from the von-Mises-Fisher distribution 
+ 
+ The emission from the source is generated following the von-Mises-Fisher distribution
+ with mean direction mu and concentration parameter kappa.
+ The sampling from the vMF distribution follows this document by Julian Straub:
+ http://people.csail.mit.edu/jstraub/download/straub2017vonMisesFisherInference.pdf
+ The emitted particles are assigned a weight so that the detected particles can be
+ reweighted to an isotropic emission distribution instead of a vMF distribution.
+ For details, see PoS (ICRC2019) 447.
+ */
+class SourceDirectedEmission: public SourceFeature {
+	Vector3d mu; // Mean emission direction in the vMF distribution
+	double kappa; // Concentration parameter of the vMF distribution
+	double ca; // helpers for the efficient calculation of frame rotation
+	double sa;
+	double cd;
+	double sd;
+public:
+	SourceDirectedEmission(Vector3d mu, double kappa);
+	void prepareCandidate(Candidate &candidate) const;
+	void setCa(double alpha);
+	void setSa(double alpha);
+	void setCd(double delta);
+	void setSd(double delta);
+	double getCa() const;
+	double getSa() const;
+	double getCd() const;
+	double getSd() const;
+	void setDescription();
+};
 
 /**
  @class SourceLambertDistributionOnSphere